Atualizado para criação dos métodos salvar, buscarPorId, deletar e atulizar.

+++ src/main/java/org/geo7/model/repository/LoteRepository.java
@@ -3,5 +3,5 @@ package org.geo7.model.repository;
 import org.geo7.model.entity.Lote;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LoteRepository extends JpaRepository<Lote, Integer> {
+public interface LoteRepository extends JpaRepository<Lote, Long> {
 }
