+++ src/main/java/org/geo7/model/repository/MunicipioRepository.java
@@ -3,5 +3,5 @@ package org.geo7.model.repository;
 import org.geo7.model.entity.Municipio;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MunicipioRepository extends JpaRepository<Municipio, Integer> {
+public interface MunicipioRepository extends JpaRepository<Municipio, Long> {
 }
