+++ src/main/java/org/geo7/model/repository/SituacaoJuridicaRepository.java
@@ -3,5 +3,5 @@ package org.geo7.model.repository;
 import org.geo7.model.entity.SituacaoJuridica;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SituacaoJuridicaRepository extends JpaRepository<SituacaoJuridica, Integer> {
+public interface SituacaoJuridicaRepository extends JpaRepository<SituacaoJuridica, Long> {
 }
